feat: move base error types from api pkg into api/apierrors
We create an alias in api/errors.go which points to the new
api/apierrors pkg so that we may slowly migrate error
helper functions to the new package as needed. Once we can
remove the aliases without compile errors we know the
migration has been completed.
