+++ internal/api/errors.go
@@ -31,43 +31,18 @@ var oauthErrorMap = map[int]string{
 	http.StatusServiceUnavailable:  "temporarily_unavailable",
 }
 
-// OAuthError is the JSON handler for OAuth2 error responses
-type OAuthError struct {
-	Err             string `json:"error"`
-	Description     string `json:"error_description,omitempty"`
-	InternalError   error  `json:"-"`
-	InternalMessage string `json:"-"`
-}
-
-func (e *OAuthError) Error() string {
-	if e.InternalMessage != "" {
-		return e.InternalMessage
-	}
-	return fmt.Sprintf("%s: %s", e.Err, e.Description)
-}
-
-// WithInternalError adds internal error information to the error
-func (e *OAuthError) WithInternalError(err error) *OAuthError {
-	e.InternalError = err
-	return e
-}
-
-// WithInternalMessage adds internal message information to the error
-func (e *OAuthError) WithInternalMessage(fmtString string, args ...interface{}) *OAuthError {
-	e.InternalMessage = fmt.Sprintf(fmtString, args...)
-	return e
-}
-
-// Cause returns the root cause error
-func (e *OAuthError) Cause() error {
-	if e.InternalError != nil {
-		return e.InternalError
-	}
-	return e
-}
+// Type aliases while we slowly refactor api errors.
+type (
+	HTTPError  = apierrors.HTTPError
+	OAuthError = apierrors.OAuthError
+)
 
 func oauthError(err string, description string) *OAuthError {
-	return &OAuthError{Err: err, Description: description}
+	return apierrors.NewOAuthError(err, description)
+}
+
+func httpError(httpStatus int, errorCode apierrors.ErrorCode, fmtString string, args ...any) *HTTPError {
+	return apierrors.NewHTTPError(httpStatus, errorCode, fmtString, args...)
 }
 
 func badRequestError(errorCode apierrors.ErrorCode, fmtString string, args ...interface{}) *HTTPError {
@@ -96,55 +71,6 @@ func tooManyRequestsError(errorCode apierrors.ErrorCode, fmtString string, args 
 
 func conflictError(fmtString string, args ...interface{}) *HTTPError {
 	return httpError(http.StatusConflict, apierrors.ErrorCodeConflict, fmtString, args...)
-}
-
-// HTTPError is an error with a message and an HTTP status code.
-type HTTPError struct {
-	HTTPStatus      int    `json:"code"`                 // do not rename the JSON tags!
-	ErrorCode       string `json:"error_code,omitempty"` // do not rename the JSON tags!
-	Message         string `json:"msg"`                  // do not rename the JSON tags!
-	InternalError   error  `json:"-"`
-	InternalMessage string `json:"-"`
-	ErrorID         string `json:"error_id,omitempty"`
-}
-
-func (e *HTTPError) Error() string {
-	if e.InternalMessage != "" {
-		return e.InternalMessage
-	}
-	return fmt.Sprintf("%d: %s", e.HTTPStatus, e.Message)
-}
-
-func (e *HTTPError) Is(target error) bool {
-	return e.Error() == target.Error()
-}
-
-// Cause returns the root cause error
-func (e *HTTPError) Cause() error {
-	if e.InternalError != nil {
-		return e.InternalError
-	}
-	return e
-}
-
-// WithInternalError adds internal error information to the error
-func (e *HTTPError) WithInternalError(err error) *HTTPError {
-	e.InternalError = err
-	return e
-}
-
-// WithInternalMessage adds internal message information to the error
-func (e *HTTPError) WithInternalMessage(fmtString string, args ...interface{}) *HTTPError {
-	e.InternalMessage = fmt.Sprintf(fmtString, args...)
-	return e
-}
-
-func httpError(httpStatus int, errorCode apierrors.ErrorCode, fmtString string, args ...interface{}) *HTTPError {
-	return &HTTPError{
-		HTTPStatus: httpStatus,
-		ErrorCode:  errorCode,
-		Message:    fmt.Sprintf(fmtString, args...),
-	}
 }
 
 // Recoverer is a middleware that recovers from panics, logs the panic (and a
